Add a practical example of using Glaive
I found it hard to use Glaive without explicit instructions like these in the
README.  I think it's also helpful to explain _why_ things have to be done this
way -- yes, the information already exists, but I think it's helpful to have it
in a more prominent place.

I'm using my own plugin as an example.  If that's too narcissistic, I'm happy to
substitute a dummy plugin.  (Then again, I believe vtd is the only maktaba
plugin other than glaive at the moment...)